Add string helper extension for generating a TF dict

+++ UniStack/Term.cs
@@ -8,6 +8,6 @@ namespace UniStack
 
         public string Value { get; set; }
 
-        public float IDF { get; set; }
+        public double IDF { get; set; }
     }
 }
